@@ -6,9 +6,11 @@ using Library.Common.Commands;
 using Library.Common.RabbitMq;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
 using Service.Activity.Handlers;
@@ -28,9 +30,9 @@ namespace Service.Activity
         public void ConfigureServices(IServiceCollection services)
         {
 
-            services.AddMvcCore();
+            services.AddControllers();
             services.AddRabbitMq(Configuration);
-            services.AddTransient<ICommandHandler<CreateActivity>, CreateActivityHandler>();
+            services.AddScoped<ICommandHandler<CreateActivity>, CreateActivityHandler>();
             services.AddSwaggerGen(c =>
             {
                 c.SwaggerDoc("v1", new OpenApiInfo { Title = "ActivityService", Version = "v1" });
@@ -38,7 +40,7 @@ namespace Service.Activity
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -46,7 +48,14 @@ namespace Service.Activity
                 app.UseSwagger();
                 app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "ActivityService v1"));
             }
-            app.UseMvc();
+
+            app.UseHttpsRedirection();
+            app.UseRouting();
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
