@@ -31,7 +31,8 @@ namespace Library.Common.Services
 
             var webHostBuilder = WebHost.CreateDefaultBuilder(args)
                 .UseConfiguration(config)
-                .UseStartup<TStartup>();
+                .UseStartup<TStartup>()
+                .UseDefaultServiceProvider(options => options.ValidateScopes = false);
 
             return new HostBuilder(webHostBuilder.Build());
         }
