@@ -1,5 +1,11 @@
-using Library.Common.Commands;
-using Library.Common.Services;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
 namespace Service.Identity
 {
@@ -7,11 +13,14 @@ namespace Service.Identity
     {
         public static void Main(string[] args)
         {
-            ServiceHost.Create<Startup>(args)
-                .UseRabbitMq()
-                .SubscribeToCommand<CreateUser>()
-                .Build()
-                .Run();
+            CreateHostBuilder(args).Build().Run();
         }
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                });
     }
 }
